+++ examples/iam/s3_account/main-datasource.tf
@@ -1,0 +1,3 @@
+data "vkcs_iam_s3_account" "s3_account" {
+  name = vkcs_iam_s3_account.s3_account.name
+}