+++ templates/iam/resources/vkcs_iam_s3_account/import.sh
@@ -1,0 +1,1 @@
+terraform import vkcs_iam_s3_account.s3_account <s3_account_id>