+++ CHANGELOG.md
@@ -6,6 +6,10 @@ description: |-
 ---
 
 # VKCS Provider's changelog
+#### v0.14.0 (unreleased)
+- Add vkcs_iam_s3_account resource
+- Add vkcs_iam_s3_account data source
+
 #### v0.13.2
 - Add details to the vkcs_compute_instance creation error
 - Add crontabs to vkcs_dataplatform_product